etl/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
quiverquant==0.2.2